+++ R/prediction_bigglm.R
@@ -1,5 +1,5 @@
-#' @rdname prediction
-#' @export
+# @rdname prediction
+# @export
 prediction.bigglm <- 
 function(model, 
          data = find_data(model, parent.frame()),