+++ R/prediction_glimQL.R
@@ -8,16 +8,18 @@ function(model,
          calculate_se = TRUE,
          ...) {
     
+    requireNamespace("aod")
+    
     type <- match.arg(type)
     
     # extract predicted values
     data <- data
     if (missing(data) || is.null(data)) {
         if (isTRUE(calculate_se)) {
-            pred <- predict(model, type = type, se.fit = TRUE, ...)
+            pred <- aod::predict(model, type = type, se.fit = TRUE, ...)
             pred <- make_data_frame(fitted = pred[["fit"]], se.fitted = pred[["se.fit"]])
         } else {
-            pred <- predict(model, type = type, se.fit = FALSE, ...)
+            pred <- aod::predict(model, type = type, se.fit = FALSE, ...)
             pred <- make_data_frame(fitted = pred, se.fitted = rep(NA_real_, length(pred)))
         }
     } else {
@@ -26,11 +28,11 @@ function(model,
         at_specification <- attr(data, "at_specification")
         # calculate predictions
         if (isTRUE(calculate_se)) {
-            tmp <- predict(model, newdata = data, type = type, se.fit = TRUE, ...)
+            tmp <- aod::predict(model, newdata = data, type = type, se.fit = TRUE, ...)
             # cbind back together
             pred <- make_data_frame(data, fitted = tmp[["fit"]], se.fitted = tmp[["se.fit"]])
         } else {
-            tmp <- predict(model, newdata = data, type = type, se.fit = FALSE, ...)
+            tmp <- aod::predict(model, newdata = data, type = type, se.fit = FALSE, ...)
             # cbind back together
             pred <- make_data_frame(data, fitted = tmp, se.fitted = rep(NA_real_, nrow(data)))
         }