+++ R/prediction_kqr.R
@@ -2,9 +2,11 @@
 #' @export
 prediction.kqr <- function(model, data, at = NULL, calculate_se = FALSE, ...) {
     
+    requireNamespace("kernlab")
+    
     # extract predicted values
     if (missing(data) || is.null(data)) {
-        pred <- make_data_frame(fitted = predict(object = model, ...)[,1L])
+        pred <- make_data_frame(fitted = kernlab::predict(object = model, ...)[,1L])
     } else {
         # setup data
         if (is.null(at)) {
@@ -13,7 +15,7 @@ prediction.kqr <- function(model, data, at = NULL, calculate_se = FALSE, ...) {
             out <- build_datalist(data, at = at, as.data.frame = TRUE)
             at_specification <- attr(out, "at_specification")
         }
-        pred <- make_data_frame(fitted = predict(model, newdata = data,...)[,1L])
+        pred <- make_data_frame(fitted = kernlab::predict(model, newdata = data,...)[,1L])
     }
     pred[["se.fitted"]] <- NA_real_