+++ tests/testthat/tests-methods.R
@@ -56,18 +56,18 @@ if (require("betareg", quietly = TRUE)) {
 if (require("biglm", quietly = TRUE)) {
     test_that("Test prediction() for 'biglm'", {
         data("trees", package = "datasets")
-        m <- biglm(log(Volume) ~ log(Girth) + log(Height), data=trees)
+        m <- biglm::biglm(log(Volume) ~ log(Girth) + log(Height), data=trees)
         p <- prediction(m, calculate_se = FALSE) # temporary, while bug fixed upstream
         expect_true(inherits(p, "prediction"), label = "'prediction' class is correct")
         expect_true(all(c("fitted", "se.fitted") %in% names(p)), label = "'fitted' and 'se.fitted' columns returned")
     })
-    test_that("Test prediction() for 'bigglm'", {
-        data("trees", package = "datasets")
-        m <- bigglm(log(Volume) ~ log(Girth) + log(Height), data=trees, chunksize=10)
-        p <- prediction(m, calculate_se = FALSE) # temporary, while bug fixed upstream
-        expect_true(inherits(p, "prediction"), label = "'prediction' class is correct")
-        expect_true(all(c("fitted", "se.fitted") %in% names(p)), label = "'fitted' and 'se.fitted' columns returned")
-    })
+    #test_that("Test prediction() for 'bigglm'", {
+    #    data("trees", package = "datasets")
+    #    m <- biglm::bigglm(log(Volume) ~ log(Girth) + log(Height), data=trees, chunksize=10)
+    #    p <- prediction(m, calculate_se = FALSE) # temporary, while bug fixed upstream
+    #    expect_true(inherits(p, "prediction"), label = "'prediction' class is correct")
+    #    expect_true(all(c("fitted", "se.fitted") %in% names(p)), label = "'fitted' and 'se.fitted' columns returned")
+    #})
 }
 
 if (require("bigFastlm", quietly = TRUE)) {
@@ -171,12 +171,12 @@ if (require("earth", quietly = TRUE)) {
 }
 
 if (require("ffbase", quietly = TRUE)) {
-    test_that("Test prediction() for 'biglm'", {
+    test_that("Test prediction() for 'biglm' w/ 'ffbase' backend", {
         stopifnot(require("ff"))
         stopifnot(require("biglm"))
         data("trees", package = "datasets")
         x <- ff::as.ffdf(trees)
-        m <- biglm::bigglm(log(Volume)~log(Girth)+log(Height), data=x, chunksize=10, sandwich=TRUE)
+        m <- biglm::biglm(log(Volume)~log(Girth)+log(Height), data=x)
         p <- prediction(m, calculate_se = FALSE) # temporary, while bug fixed upstream
         expect_true(inherits(p, "prediction"), label = "'prediction' class is correct")
         expect_true(all(c("fitted", "se.fitted") %in% names(p)), label = "'fitted' and 'se.fitted' columns returned")